backend/pyproject.toml
[project]
name = "oneidp"
version = "1.0.0"
description = "基于 Onebot-V11 协议的 IDP 提供者，帮助 QQ 群群友绑定 SSO 账号并完成登录授权"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "OneIDP Team"}
]

dependencies = [
    # Web 框架
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    
    # 数据库
    "sqlalchemy[asyncio]>=2.0.25",
    "aiosqlite>=0.19.0",  # SQLite 异步驱动
    "asyncpg>=0.29.0",  # PostgreSQL 异步驱动 (PostgreSQL 14+)
    
    # 机器人通讯
    "aiocqhttp>=1.4.4",
    "websockets>=12.0",
    
    # HTTP 客户端
    "httpx>=0.26.0",
    
    # 配置管理
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    
    # 安全相关
    "pyjwt>=2.8.0",  # JWT 令牌
    
    # 工具库
    "python-multipart>=0.0.6",  # 表单解析
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

[project.scripts]
oneidp = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build.targets.sdist]
include = ["*.py", "bot/", "database/", "oauth/", "page/", "utils/"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "SIM", "ARG"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
